Fully drop Python 2 support
