+++ napalm_logs/device.py
@@ -22,7 +22,6 @@ from prometheus_client import Counter
 
 # Import napalm-logs pkgs
 import napalm_logs.utils
-import napalm_logs.ext.six as six
 from napalm_logs.proc import NapalmLogsProc
 from napalm_logs.config import PUB_PX_IPC_URL
 from napalm_logs.config import DEV_IPC_URL
